move telescope
